+++ docs/conf.py
@@ -71,9 +71,9 @@ copyright = '2015-2017 — Mozilla Services'
 # built documents.
 #
 # The short X.Y version.
-version = '6.0'
+version = '6.1.0'
 # The full version, including alpha/beta/rc tags.
-release = '6.0.9'
+release = '6.1.0'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.