+++ CHANGELOG.rst
@@ -10,6 +10,10 @@ This document describes changes between each past release.
 
 - ``delete()`` method from cache backend now returns the deleted value (fixes #1231)
 
+**Bug fixes**
+
+- Do not allow to reuse deletion tokens (fixes #1171)
+
 
 7.0.1 (2017-05-17)
 ------------------
@@ -95,7 +99,6 @@ Protocol is now at version **1.16**. See `API changelog`_.
   other buckets and collections when deleting a bucket. (fixes #1209)
 - Fix race conditions on deletions with upsert in PostgreSQL ``storage.update()`` (fixes #1202)
 - Fix PostgreSQL backend race condition when replacing permissions of an object (fixes #1210)
-- Fix crash when deleting multiple buckets with quotas plugin enabled (fixes #1201)
 
 **Internal changes**