+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ This document describes changes between each past release.
   events for buckets and collections that already exist. This causes
   the ``quotas`` plugin to no longer leak "quota" when used with the
   ``default_bucket`` plugin. (#1226)
+- Do not allow to reuse deletion tokens (fixes #1171)
 
 
 7.0.1 (2017-05-17)
@@ -102,7 +103,6 @@ Protocol is now at version **1.16**. See `API changelog`_.
   other buckets and collections when deleting a bucket. (fixes #1209)
 - Fix race conditions on deletions with upsert in PostgreSQL ``storage.update()`` (fixes #1202)
 - Fix PostgreSQL backend race condition when replacing permissions of an object (fixes #1210)
-- Fix crash when deleting multiple buckets with quotas plugin enabled (fixes #1201)
 
 **Internal changes**