+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 
 This document describes changes between each past release.
 
-6.0.9 (2017-05-25)
+6.1.0 (2017-05-25)
 ------------------
 
 **New feature**